Adding docs while I try to understand how to make 1 steal 1 agent. Moved some implementation details from steal into separate modules.
