+++ mirrord/agent/src/steal.rs
@@ -4,17 +4,11 @@ use std::{
     path::PathBuf,
 };
 
-use futures::{
-    future::{FutureExt, OptionFuture},
-    stream,
-    stream::TryStreamExt,
-    StreamExt,
-};
+use futures::StreamExt;
 use mirrord_protocol::{
     tcp::{DaemonTcp, LayerTcpSteal, NewTcpConnection, TcpClose, TcpData},
     ConnectionId, Port,
 };
-use rand::distributions::{Alphanumeric, DistString};
 use streammap_ext::StreamMap;
 use tokio::{
     io::{AsyncWriteExt, ReadHalf, WriteHalf},
@@ -25,187 +19,51 @@ use tokio::{
 use tokio_util::{io::ReaderStream, sync::CancellationToken};
 use tracing::{debug, error, info, log::warn};
 
+use self::ip_tables::SafeIpTables;
 use crate::{
     error::{AgentError, Result},
     runtime::set_namespace,
     util::{ClientID, IndexAllocator, Subscriptions},
 };
 
-#[cfg_attr(test, mockall::automock)]
-trait IPTables {
-    fn create_chain(&self, name: &str) -> Result<()>;
-    fn remove_chain(&self, name: &str) -> Result<()>;
+mod ip_tables;
+mod orig_dst;
 
-    fn add_rule(&self, chain: &str, rule: &str) -> Result<()>;
-    fn insert_rule(&self, chain: &str, rule: &str, index: i32) -> Result<()>;
-    fn list_rules(&self, chain: &str) -> Result<Vec<String>>;
-    fn remove_rule(&self, chain: &str, rule: &str) -> Result<()>;
-}
-
-impl IPTables for iptables::IPTables {
-    fn create_chain(&self, name: &str) -> Result<()> {
-        self.new_chain(IPTABLES_TABLE_NAME, name)
-            .map_err(|e| AgentError::IPTablesError(e.to_string()))?;
-        self.append(IPTABLES_TABLE_NAME, name, "-j RETURN")
-            .map_err(|e| AgentError::IPTablesError(e.to_string()))?;
-
-        Ok(())
-    }
-
-    fn remove_chain(&self, name: &str) -> Result<()> {
-        self.flush_chain(IPTABLES_TABLE_NAME, name)
-            .map_err(|e| AgentError::IPTablesError(e.to_string()))?;
-        self.delete_chain(IPTABLES_TABLE_NAME, name)
-            .map_err(|e| AgentError::IPTablesError(e.to_string()))?;
-
-        Ok(())
-    }
-
-    fn add_rule(&self, chain: &str, rule: &str) -> Result<()> {
-        self.append(IPTABLES_TABLE_NAME, chain, rule)
-            .map_err(|e| AgentError::IPTablesError(e.to_string()))
-    }
-
-    fn insert_rule(&self, chain: &str, rule: &str, index: i32) -> Result<()> {
-        self.insert(IPTABLES_TABLE_NAME, chain, rule, index)
-            .map_err(|e| AgentError::IPTablesError(e.to_string()))
-    }
-
-    fn list_rules(&self, chain: &str) -> Result<Vec<String>> {
-        self.list(IPTABLES_TABLE_NAME, chain)
-            .map_err(|e| AgentError::IPTablesError(e.to_string()))
-    }
-
-    fn remove_rule(&self, chain: &str, rule: &str) -> Result<()> {
-        self.delete(IPTABLES_TABLE_NAME, chain, rule)
-            .map_err(|e| AgentError::IPTablesError(e.to_string()))
-    }
-}
-
-/// Wrapper struct for IPTables so it flushes on drop.
-struct SafeIpTables<IPT: IPTables> {
-    inner: IPT,
-    chain_name: String,
-    formatter: IPTableFormatter,
-}
-
-const IPTABLES_TABLE_NAME: &str = "nat";
-/// Wrapper for using iptables. This creates a a new chain on creation and deletes it on drop.
-/// The way it works is that it adds a chain, then adds a rule to the chain that returns to the
-/// original chain (fallback) and adds a rule in the "PREROUTING" table that jumps to the new chain.
-/// Connections will go then PREROUTING -> OUR_CHAIN -> IF MATCH REDIRECT -> IF NOT MATCH FALLBACK
-/// -> ORIGINAL_CHAIN
-impl<IPT> SafeIpTables<IPT>
-where
-    IPT: IPTables,
-{
-    pub fn new(ipt: IPT) -> Result<Self> {
-        let formatter = IPTableFormatter::detect(&ipt)?;
-
-        let random_string = Alphanumeric.sample_string(&mut rand::thread_rng(), 5);
-        let chain_name = format!("MIRRORD_REDIRECT_{}", random_string);
-
-        ipt.create_chain(&chain_name)?;
-
-        ipt.add_rule(formatter.entrypoint(), &format!("-j {}", chain_name))?;
-
-        Ok(Self {
-            inner: ipt,
-            chain_name,
-            formatter,
-        })
-    }
-
-    #[tracing::instrument(level = "trace", skip(self))]
-    pub fn add_redirect(&self, redirected_port: Port, target_port: Port) -> Result<()> {
-        self.inner.insert_rule(
-            &self.chain_name,
-            &self.formatter.redirect_rule(redirected_port, target_port),
-            1,
-        )
-    }
-
-    pub fn remove_redirect(&self, redirected_port: Port, target_port: Port) -> Result<()> {
-        self.inner.remove_rule(
-            &self.chain_name,
-            &self.formatter.redirect_rule(redirected_port, target_port),
-        )
-    }
-}
-
-impl<IPT> Drop for SafeIpTables<IPT>
-where
-    IPT: IPTables,
-{
-    fn drop(&mut self) {
-        self.inner
-            .remove_rule(
-                self.formatter.entrypoint(),
-                &format!("-j {}", self.chain_name),
-            )
-            .unwrap();
-
-        self.inner.remove_chain(&self.chain_name).unwrap();
-    }
-}
-
-enum IPTableFormatter {
-    Normal,
-    Mesh,
-}
-
-impl IPTableFormatter {
-    const MESH_OUTPUTS: [&'static str; 2] = ["-j PROXY_INIT_OUTPUT", "-j ISTIO_OUTPUT"];
-
-    fn detect<IPT: IPTables>(ipt: &IPT) -> Result<Self> {
-        let output = ipt.list_rules("OUTPUT")?;
-
-        if output.iter().any(|rule| {
-            IPTableFormatter::MESH_OUTPUTS
-                .iter()
-                .any(|mesh_output| rule.contains(mesh_output))
-        }) {
-            Ok(IPTableFormatter::Mesh)
-        } else {
-            Ok(IPTableFormatter::Normal)
-        }
-    }
-
-    fn entrypoint(&self) -> &str {
-        match self {
-            IPTableFormatter::Normal => "PREROUTING",
-            IPTableFormatter::Mesh => "OUTPUT",
-        }
-    }
-
-    fn redirect_rule(&self, redirected_port: Port, target_port: Port) -> String {
-        let redirect_rule = format!(
-            "-m tcp -p tcp --dport {} -j REDIRECT --to-ports {}",
-            redirected_port, target_port
-        );
-
-        match self {
-            IPTableFormatter::Normal => redirect_rule,
-            IPTableFormatter::Mesh => {
-                format!("-o lo {}", redirect_rule)
-            }
-        }
-    }
-}
-
+/// Commands from the agent that are passed down to the stealer worker, through [`TcpStealerAPI`].
+///
+/// These are the operations that the agent receives from the layer to make the _steal_ feature
+/// work.
 #[derive(Debug)]
-enum StealerCommands {
+enum Command {
+    /// Passes the channel that's used by the worker to respond back to the (agent -> layer).
     NewAgent(Sender<DaemonTcp>),
+
+    /// A layer wants to subscribe to this [`Port`].
+    ///
+    /// The agent starts stealing traffic on this [`Port`].
     Subscribe(Port),
+
+    /// A layer wants to unsubscribe from this [`Port`].
+    ///
+    /// The agent stops stealing traffic from this [`Port`].
     UnsubscribePort(Port),
-    UnsubscribeConnection(ConnectionId),
+
+    /// Part of the [`Drop`] implementation of [`TcpStealerAPI`].
+    ///
+    /// Closes a layer connection, and unsubscribe its ports.
     AgentClosed,
 }
 
+/// Association between a client (identified by the `client_id`) and a [`Command`].
+///
+/// The (agent -> worker) channel uses this, instead of naked [`Command`]s when communicating.
 #[derive(Debug)]
 pub struct StealerCommand {
+    /// Identifies which layer instance is sending the [`Command`].
     client_id: ClientID,
-    command: StealerCommands,
+
+    /// The command message sent from (layer -> agent) to be handled by the steal worker.
+    command: Command,
 }
 
 #[derive(Debug)]
@@ -217,14 +75,23 @@ pub(super) struct TcpStealerAPI {
 }
 
 /// Created once per agent during initialization.
+///
+/// Runs as a separate thread while the agent lives.
+///
+/// - (agent -> stealer) communication is handled by [`command_rx`];
+/// - (stealer -> agent) communication is handled by [`client_senders`], and the [`Sender`] channels
+///   come inside [`StealerCommand`]s through  [`command_rx`];
 #[derive(Debug)]
 pub(super) struct TcpConnectionStealer {
     port_subscriptions: Subscriptions<Port, ClientID>,
 
-    /// Communication between the agent and the stealer task.
+    /// Communication between (agent -> stealer) task.
     ///
     /// The agent controls the stealer task through [`TcpStealerAPI::command_tx`].
     command_rx: Receiver<StealerCommand>,
+
+    /// Connected clients (layer instances) and the channels which the stealer task uses to send
+    /// back messages (stealer -> agent -> layer).
     client_senders: HashMap<ClientID, Sender<DaemonTcp>>,
     index_allocator: IndexAllocator<ConnectionId>,
 }
@@ -255,7 +122,7 @@ impl TcpConnectionStealer {
     /// Runs the stealer loop.
     #[tracing::instrument(level = "trace", skip(self))]
     pub(super) async fn start(
-        self,
+        mut self,
         cancellation_token: CancellationToken,
     ) -> Result<(), AgentError> {
         loop {
@@ -266,10 +133,11 @@ impl TcpConnectionStealer {
                     } else { break; }
                 },
                 // TODO(alex) [mid] 2022-12-01: This should be global as well?
+                //
                 // Like steal everything or only steal what the users asked for?
-                packet = self.raw_capture.next() => {
-                    self.handle_packet(packet?).await?;
-                }
+                //
+                // If we do this, then we break stuff, as this would mean we take every `TcpStream`.
+                // accept = listener.accept() => todo!()
                 _ = cancellation_token.cancelled() => {
                     break;
                 }
@@ -288,7 +156,7 @@ impl TcpConnectionStealer {
     async fn subscribe(&mut self, client_id: ClientID, port: Port) -> Result<(), AgentError> {
         self.port_subscriptions.subscribe(client_id, port);
         // self.update_stealer()?;
-        self.send_message_to_client(&client_id, DaemonTcp::Subscribed)
+        self.send_message_to_single_client(&client_id, DaemonTcp::Subscribed)
             .await
     }
 
@@ -303,7 +171,7 @@ impl TcpConnectionStealer {
 
     /// Sends a [`DaemonTcp`] message back to the client with `client_id`.
     // #[tracing::instrument(level = "trace", skip(self))]
-    async fn send_message_to_client(
+    async fn send_message_to_single_client(
         &mut self,
         client_id: &ClientID,
         message: DaemonTcp,
@@ -323,15 +191,29 @@ impl TcpConnectionStealer {
     }
 
     // #[tracing::instrument(level = "trace", skip(self, clients))]
-    async fn broadcast_message(
+    async fn send_message_to_all_clients(
         &mut self,
         clients: impl Iterator<Item = &ClientID>,
         message: DaemonTcp,
     ) -> Result<(), AgentError> {
         for client_id in clients {
-            self.send_message_to_client(client_id, message.clone())
+            self.send_message_to_single_client(client_id, message.clone())
                 .await?;
         }
+        Ok(())
+    }
+
+    // #[tracing::instrument(level = "trace", skip(self))]
+    async fn handle_command(&mut self, command: StealerCommand) -> Result<(), AgentError> {
+        let StealerCommand { client_id, command } = command;
+
+        match command {
+            Command::NewAgent(daemon_tx) => todo!(),
+            Command::Subscribe(port) => todo!(),
+            Command::UnsubscribePort(port) => todo!(),
+            Command::AgentClosed => todo!(),
+        }
+
         Ok(())
     }
 }
@@ -507,245 +389,4 @@ pub async fn steal_worker(
     let listen_port = listener.local_addr()?.port();
 
     StealWorker::new(tx, listen_port)?.start(rx, listener).await
-}
-
-// orig_dst borrowed from linkerd2-proxy
-// https://github.com/linkerd/linkerd2-proxy/blob/main/linkerd/proxy/transport/src/orig_dst.rs
-// copyright 2018 the linkerd2-proxy authors
-mod orig_dst {
-    use std::{io, net::SocketAddr};
-
-    use tokio::net::TcpStream;
-
-    #[cfg(target_os = "linux")]
-    #[allow(unsafe_code)]
-    #[tracing::instrument(level = "trace")]
-    pub fn orig_dst_addr(sock: &TcpStream) -> io::Result<SocketAddr> {
-        use std::os::unix::io::AsRawFd;
-        let fd = sock.as_raw_fd();
-        unsafe { linux::so_original_dst(fd) }
-    }
-
-    #[cfg(not(target_os = "linux"))]
-    pub fn orig_dst_addr(_: &TcpStream) -> io::Result<SocketAddr> {
-        Err(io::Error::new(
-            io::ErrorKind::Other,
-            "SO_ORIGINAL_DST not supported on this operating system",
-        ))
-    }
-
-    #[cfg(target_os = "linux")]
-    #[allow(unsafe_code)]
-    mod linux {
-        use std::{
-            io, mem,
-            net::{Ipv4Addr, Ipv6Addr, SocketAddr, SocketAddrV4, SocketAddrV6},
-            os::unix::io::RawFd,
-        };
-
-        use tracing::warn;
-
-        pub unsafe fn so_original_dst(fd: RawFd) -> io::Result<SocketAddr> {
-            let mut sockaddr: libc::sockaddr_storage = mem::zeroed();
-            let mut socklen: libc::socklen_t = mem::size_of::<libc::sockaddr_storage>() as u32;
-
-            let ret = libc::getsockopt(
-                fd,
-                libc::SOL_IP,
-                libc::SO_ORIGINAL_DST,
-                &mut sockaddr as *mut _ as *mut _,
-                &mut socklen as *mut _ as *mut _,
-            );
-            if ret != 0 {
-                let e = io::Error::last_os_error();
-                warn!("failed to read SO_ORIGINAL_DST: {:?}", e);
-                return Err(e);
-            }
-
-            mk_addr(&sockaddr, socklen)
-        }
-
-        // Borrowed with love from net2-rs
-        // https://github.com/rust-lang-nursery/net2-rs/blob/1b4cb4fb05fbad750b271f38221eab583b666e5e/src/socket.rs#L103
-        //
-        // Copyright (c) 2014 The Rust Project Developers
-        fn mk_addr(
-            storage: &libc::sockaddr_storage,
-            len: libc::socklen_t,
-        ) -> io::Result<SocketAddr> {
-            match storage.ss_family as libc::c_int {
-                libc::AF_INET => {
-                    assert!(len as usize >= mem::size_of::<libc::sockaddr_in>());
-
-                    let sa = {
-                        let sa = storage as *const _ as *const libc::sockaddr_in;
-                        unsafe { *sa }
-                    };
-
-                    let bits = ntoh32(sa.sin_addr.s_addr);
-                    let ip = Ipv4Addr::new(
-                        (bits >> 24) as u8,
-                        (bits >> 16) as u8,
-                        (bits >> 8) as u8,
-                        bits as u8,
-                    );
-                    let port = sa.sin_port;
-                    Ok(SocketAddr::V4(SocketAddrV4::new(ip, ntoh16(port))))
-                }
-                libc::AF_INET6 => {
-                    assert!(len as usize >= mem::size_of::<libc::sockaddr_in6>());
-
-                    let sa = {
-                        let sa = storage as *const _ as *const libc::sockaddr_in6;
-                        unsafe { *sa }
-                    };
-
-                    let arr = sa.sin6_addr.s6_addr;
-                    let ip = Ipv6Addr::new(
-                        (arr[0] as u16) << 8 | (arr[1] as u16),
-                        (arr[2] as u16) << 8 | (arr[3] as u16),
-                        (arr[4] as u16) << 8 | (arr[5] as u16),
-                        (arr[6] as u16) << 8 | (arr[7] as u16),
-                        (arr[8] as u16) << 8 | (arr[9] as u16),
-                        (arr[10] as u16) << 8 | (arr[11] as u16),
-                        (arr[12] as u16) << 8 | (arr[13] as u16),
-                        (arr[14] as u16) << 8 | (arr[15] as u16),
-                    );
-
-                    let port = sa.sin6_port;
-                    let flowinfo = sa.sin6_flowinfo;
-                    let scope_id = sa.sin6_scope_id;
-                    Ok(SocketAddr::V6(SocketAddrV6::new(
-                        ip,
-                        ntoh16(port),
-                        flowinfo,
-                        scope_id,
-                    )))
-                }
-                _ => Err(io::Error::new(
-                    io::ErrorKind::InvalidInput,
-                    "invalid argument",
-                )),
-            }
-        }
-
-        fn ntoh16(i: u16) -> u16 {
-            <u16>::from_be(i)
-        }
-
-        fn ntoh32(i: u32) -> u32 {
-            <u32>::from_be(i)
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use mockall::predicate::*;
-
-    use super::*;
-
-    #[test]
-    fn default() {
-        let mut mock = MockIPTables::new();
-
-        mock.expect_list_rules()
-            .with(eq("OUTPUT"))
-            .returning(|_| Ok(vec![]));
-
-        mock.expect_create_chain()
-            .with(str::starts_with("MIRRORD_REDIRECT_"))
-            .times(1)
-            .returning(|_| Ok(()));
-
-        mock.expect_remove_chain()
-            .with(str::starts_with("MIRRORD_REDIRECT_"))
-            .times(1)
-            .returning(|_| Ok(()));
-
-        mock.expect_add_rule()
-            .with(eq("PREROUTING"), str::starts_with("-j MIRRORD_REDIRECT_"))
-            .times(1)
-            .returning(|_, _| Ok(()));
-
-        mock.expect_insert_rule()
-            .with(
-                str::starts_with("MIRRORD_REDIRECT_"),
-                eq("-m tcp -p tcp --dport 69 -j REDIRECT --to-ports 420"),
-                eq(1),
-            )
-            .times(1)
-            .returning(|_, _, _| Ok(()));
-
-        mock.expect_remove_rule()
-            .with(eq("PREROUTING"), str::starts_with("-j MIRRORD_REDIRECT_"))
-            .times(1)
-            .returning(|_, _| Ok(()));
-
-        mock.expect_remove_rule()
-            .with(
-                str::starts_with("MIRRORD_REDIRECT_"),
-                eq("-m tcp -p tcp --dport 69 -j REDIRECT --to-ports 420"),
-            )
-            .times(1)
-            .returning(|_, _| Ok(()));
-
-        let ipt = SafeIpTables::new(mock).expect("Create Failed");
-
-        assert!(ipt.add_redirect(69, 420).is_ok());
-
-        assert!(ipt.remove_redirect(69, 420).is_ok());
-    }
-
-    #[test]
-    fn linkerd() {
-        let mut mock = MockIPTables::new();
-
-        mock.expect_list_rules()
-            .with(eq("OUTPUT"))
-            .returning(|_| Ok(vec!["-j PROXY_INIT_OUTPUT".to_owned()]));
-
-        mock.expect_create_chain()
-            .with(str::starts_with("MIRRORD_REDIRECT_"))
-            .times(1)
-            .returning(|_| Ok(()));
-
-        mock.expect_remove_chain()
-            .with(str::starts_with("MIRRORD_REDIRECT_"))
-            .times(1)
-            .returning(|_| Ok(()));
-
-        mock.expect_add_rule()
-            .with(eq("OUTPUT"), str::starts_with("-j MIRRORD_REDIRECT_"))
-            .times(1)
-            .returning(|_, _| Ok(()));
-
-        mock.expect_insert_rule()
-            .with(
-                str::starts_with("MIRRORD_REDIRECT_"),
-                eq("-o lo -m tcp -p tcp --dport 69 -j REDIRECT --to-ports 420"),
-                eq(1),
-            )
-            .times(1)
-            .returning(|_, _, _| Ok(()));
-
-        mock.expect_remove_rule()
-            .with(eq("OUTPUT"), str::starts_with("-j MIRRORD_REDIRECT_"))
-            .times(1)
-            .returning(|_, _| Ok(()));
-
-        mock.expect_remove_rule()
-            .with(
-                str::starts_with("MIRRORD_REDIRECT_"),
-                eq("-o lo -m tcp -p tcp --dport 69 -j REDIRECT --to-ports 420"),
-            )
-            .times(1)
-            .returning(|_, _| Ok(()));
-
-        let ipt = SafeIpTables::new(mock).expect("Create Failed");
-
-        assert!(ipt.add_redirect(69, 420).is_ok());
-
-        assert!(ipt.remove_redirect(69, 420).is_ok());
-    }
 }
